chore: Rename hcl-helpers to hcl-tools
