+++ packages/@cdktf/cli-core/package.json
@@ -42,7 +42,7 @@
     "@cdktf/hcl2json": "0.0.0",
     "@cdktf/provider-schema": "0.0.0",
     "@cdktf/node-pty-prebuilt-multiarch": "0.10.1-pre.11",
-    "@cdktf/hcl-helpers": "0.0.0",
+    "@cdktf/hcl-tools": "0.0.0",
     "@sentry/node": "^7.80.0",
     "archiver": "^5.3.2",
     "cdktf": "0.0.0",
