--- packages/@cdktf/hcl-helpers/README.md
@@ -1,11 +1,0 @@
-# `@cdktf/hcl-helpers`
-
-> TODO: description
-
-## Usage
-
-```
-const hclHelpers = require('@cdktf/hcl-helpers');
-
-// TODO: DEMONSTRATE API
-```
