--- packages/@cdktf/hcl-helpers/go.mod
@@ -1,5 +1,0 @@
-module github.com/hashicorp/terraform-cdk/hcl-helpers
-
-go 1.16
-
-require github.com/hashicorp/hcl/v2 v2.19.1
