--- packages/@cdktf/hcl-helpers/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@cdktf/hcl-helpers",
-  "version": "0.0.0",
-  "description": "Interact with HCL helper libraries for output of HCL configuration",
-  "keywords": [
-    "cdk",
-    "cdktf",
-    "terraform",
-    "hcl"
-  ],
-  "author": "Mutahhir Hayat <mutahhir.hayat@hashicorp.com>",
-  "homepage": "https://cdk.tf",
-  "license": "MPL-2.0",
-  "main": "src/index.js",
-  "scripts": {
-    "build": "tsc && ./build-go.sh",
-    "watch": "tsc -w",
-    "watch-preserve-output": "tsc -w --preserveWatchOutput",
-    "test": "jest",
-    "test:ci": "jest --ci",
-    "jest-watch": "jest --watch",
-    "package": "./package.sh",
-    "dist-clean": "rm -rf dist",
-    "prebuild": "./prebuild.sh"
-  },
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/hashicorp/terraform-cdk.git",
-    "directory": "packages/@cdktf/hcl-helpers"
-  },
-  "bugs": {
-    "url": "https://github.com/hashicorp/terraform-cdk/issues"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.10",
-    "@types/node": "^20.9.4",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.3.2"
-  }
-}
