--- packages/@cdktf/hcl-helpers/src/index.ts
@@ -1,6 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-export { format } from "./bridge";
