--- packages/@cdktf/hcl-helpers/tsconfig.json
@@ -1,34 +1,0 @@
-{
-  "compilerOptions": {
-    "alwaysStrict": true,
-    "baseUrl": ".",
-    "declaration": true,
-    "declarationMap": true,
-    "experimentalDecorators": true,
-    "inlineSourceMap": true,
-    "inlineSources": true,
-    "lib": ["es2019"],
-    "module": "CommonJS",
-    "noEmitOnError": false,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "jsx": "react",
-    "resolveJsonModule": true,
-    "strict": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
-    "skipLibCheck": true,
-    "stripInternal": true,
-    "target": "ES2018",
-    "incremental": true,
-    "esModuleInterop": true,
-    "allowJs": true,
-    "checkJs": false
-  },
-  "include": ["src/**/*.ts", "test/**/*.ts"],
-  "exclude": ["node_modules", "**/*.d.ts", "wasm/*"]
-}
