--- packages/@cdktf/hcl-helpers/wasm/.gitignore
@@ -1,4 +1,0 @@
-wasm_exec.js
-!wasm_exec.d.ts
-!bridge_wasm_exec.js
-!bridge_wasm_exec.d.ts
