--- packages/@cdktf/hcl-helpers/wasm/.npmignore
@@ -1,3 +1,0 @@
-!wasm_exec.js
-!bridge_wasm_exec.js
-!wasm_exec.d.ts
