--- packages/@cdktf/hcl-helpers/wasm/bridge_wasm_exec.d.ts
@@ -1,3 +1,0 @@
-// Copyright (c) HashiCorp, Inc
-// SPDX-License-Identifier: MPL-2.0
-export {};
