--- packages/@cdktf/hcl-helpers/wasm/wasm_exec.d.ts
@@ -1,3 +1,0 @@
-// Copyright (c) HashiCorp, Inc
-// Licensed under the Mozilla Public License 2.0
-export {};
