+++ packages/@cdktf/hcl-tools/go.mod
@@ -1,0 +1,5 @@
+module github.com/hashicorp/terraform-cdk/hcl-tools
+
+go 1.16
+
+require github.com/hashicorp/hcl/v2 v2.19.1
