+++ packages/@cdktf/hcl-tools/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+export { format } from "./bridge";
