+++ packages/@cdktf/hcl-tools/wasm/.gitignore
@@ -1,0 +1,4 @@
+wasm_exec.js
+!wasm_exec.d.ts
+!bridge_wasm_exec.js
+!bridge_wasm_exec.d.ts
