+++ packages/@cdktf/hcl-tools/wasm/.npmignore
@@ -1,0 +1,3 @@
+!wasm_exec.js
+!bridge_wasm_exec.js
+!wasm_exec.d.ts
