+++ packages/@cdktf/hcl-tools/wasm/bridge_wasm_exec.d.ts
@@ -1,0 +1,3 @@
+// Copyright (c) HashiCorp, Inc
+// SPDX-License-Identifier: MPL-2.0
+export {};
