+++ packages/@cdktf/hcl-tools/wasm/wasm_exec.d.ts
@@ -1,0 +1,3 @@
+// Copyright (c) HashiCorp, Inc
+// Licensed under the Mozilla Public License 2.0
+export {};
