+++ packages/cdktf-cli/build.ts
@@ -73,7 +73,7 @@ const nativeNodeModulesPlugin = {
       "jsii",
       "jsii-pacmak",
       "@cdktf/hcl2json",
-      "@cdktf/hcl-helpers",
+      "@cdktf/hcl-tools",
       "yargs",
       "cdktf",
       "@cdktf/hcl2cdk",
